+++ lead/src/main/java/com/ironhack/cmr/lead/repository/LeadRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LeadRepository extends JpaRepository<Lead, Integer> {
+public interface LeadRepository extends JpaRepository<Lead, Long> {
 }
